requirements.txt
# Core Dependencies
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
plotly>=5.15.0
opencv-python>=4.8.0
Pillow>=9.5.0

# OCR Engines
easyocr>=1.7.0
pytesseract>=0.3.10

# Deep Learning (for EasyOCR)
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
scikit-image>=0.21.0

# Web Framework
requests>=2.27.0

# Environment Management
python-dotenv>=1.0.0

# Development Dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0